fix issue with crash when running /say in a private chat

+++ app/commands/say.js
@@ -1,6 +1,8 @@
 module.exports = () => ctx => {
-  ctx.deleteMessage(ctx.message.id);
+  // messages in private chats can't be deleted
+  if (ctx.chat.type !== 'private') ctx.deleteMessage(ctx.message.id);
 
   let msg = ctx.state.command.args;
+  // prevent bot from sending empty messages
   if (msg) ctx.reply(msg);
 };
